grafana_viewer/requirements.txt
flask
requests
PyYAML
python-dotenv
